[nrf5] Put common code for examples in seperate directory (#1416)
* [nrf5] Put common code for examples in seperate directory

* fix markdown format

* Remove extern "C" since main function lies in a CPP source

* examples/common/nrf5 -> examples/platform/nrf528xx
