+++ examples/common/nrf5/README.md
@@ -1,0 +1,16 @@
+# CHIP nRF52840 Common Code
+
+This directory contains common code for several example apps.
+
+-   `common`
+
+    This directory contains common code for bringing up a nRF52840 example app.
+
+-   `util`
+
+    This directory contains some usful classes for example apps.
+
+-   `gdb`
+
+    This directory contains some shell scripts for setting up debugging
+    environment.
