+++ examples/common/nrf5/gdb/gdb-startup-cmds.txt
