+++ examples/common/nrf5/gdb/start-gdb.sh
