+++ examples/common/nrf5/gdb/start-jlink-gdb-server.sh
