+++ examples/lock-app/nrf5/Makefile
@@ -17,7 +17,7 @@
 #   limitations under the License.
 #
 
-# 
+#
 #   @file
 #         Makefile for building the CHIP nRF52840 Lock Example Application.
 #
@@ -26,6 +26,7 @@ PROJECT_ROOT := $(realpath .)
 
 CHIP_ROOT ?= $(realpath $(PROJECT_ROOT)/third_party/connectedhomeip)
 BUILD_SUPPORT_DIR = $(CHIP_ROOT)/config/nrf5
+NRF5_PLATFORM_DIR = $(CHIP_ROOT)/examples/platform/nrf528xx
 
 include $(BUILD_SUPPORT_DIR)/nrf5-app.mk
 include $(BUILD_SUPPORT_DIR)/nrf5-chip.mk
@@ -35,13 +36,14 @@ APP = chip-nrf52840-lock-example
 SRCS = \
     $(PROJECT_ROOT)/main/main.cpp \
     $(PROJECT_ROOT)/main/AppTask.cpp \
-    $(PROJECT_ROOT)/main/LEDWidget.cpp \
     $(PROJECT_ROOT)/main/BoltLockManager.cpp \
-    $(PROJECT_ROOT)/main/Server.cpp \
     $(PROJECT_ROOT)/main/DataModelHandler.cpp \
-    $(PROJECT_ROOT)/main/support/CXXExceptionStubs.cpp \
-    $(PROJECT_ROOT)/main/support/nRF5Sbrk.c \
-    $(PROJECT_ROOT)/main/support/FreeRTOSNewlibLockSupport.c \
+    $(NRF5_PLATFORM_DIR)/app/support/CXXExceptionStubs.cpp \
+    $(NRF5_PLATFORM_DIR)/app/support/nRF5Sbrk.c \
+    $(NRF5_PLATFORM_DIR)/app/support/FreeRTOSNewlibLockSupport.c \
+    $(NRF5_PLATFORM_DIR)/app/Server.cpp \
+    $(NRF5_PLATFORM_DIR)/app/chipinit.cpp \
+    $(NRF5_PLATFORM_DIR)/util/LEDWidget.cpp \
     $(NRF5_SDK_ROOT)/components/ble/common/ble_advdata.c \
     $(NRF5_SDK_ROOT)/components/ble/common/ble_srv_common.c \
     $(NRF5_SDK_ROOT)/components/ble/nrf_ble_gatt/nrf_ble_gatt.c \
@@ -115,6 +117,9 @@ INC_DIRS = \
     $(CHIP_ROOT)/src/lib \
     $(CHIP_ROOT)/src/ \
     $(CHIP_ROOT)/src/system \
+    $(NRF5_PLATFORM_DIR)/app/include \
+    $(NRF5_PLATFORM_DIR)/app/project_include \
+    $(NRF5_PLATFORM_DIR)/util/include \
     $(NRF5_SDK_ROOT)/components \
     $(NRF5_SDK_ROOT)/components/boards \
     $(NRF5_SDK_ROOT)/components/ble/ble_advertising \
@@ -272,8 +277,8 @@ else
     CHIP_DEFINES += CHIP_CONFIG_LOG_LEVEL=OT_LOG_LEVEL_INFO
 endif
 
-CHIP_PROJECT_CONFIG = $(PROJECT_ROOT)/main/include/CHIPProjectConfig.h
+CHIP_PROJECT_CONFIG = $(NRF5_PLATFORM_DIR)/app/project_include/CHIPProjectConfig.h
 
-LINKER_SCRIPT = $(PROJECT_ROOT)/main/ldscripts/chip-nrf52840-lock-example.ld
+LINKER_SCRIPT = $(NRF5_PLATFORM_DIR)/app/ldscripts/chip-nrf52840-example.ld
 
 $(call GenerateBuildRules)
