+++ examples/platform/nrf528xx/app/include/chipinit.h
@@ -1,0 +1,8 @@
+#ifndef NRF5_CHIPINIT_H
+#define NRF5_CHIPINIT_H
+
+#include "sdk_common.h"
+
+ret_code_t ChipInit();
+
+#endif // NRF5_CHIPINIT_H
