+++ examples/platform/nrf528xx/app/ldscripts/chip-nrf52840-example.ld
