+++ examples/platform/nrf528xx/app/project_include/CHIPProjectConfig.h
