+++ examples/platform/nrf528xx/app/project_include/OpenThreadConfig.h
