+++ examples/platform/nrf528xx/app/project_include/nrf_log_ctrl_internal.h
