+++ examples/platform/nrf528xx/app/support/CXXExceptionStubs.cpp
