+++ examples/platform/nrf528xx/app/support/FreeRTOSNewlibLockSupport.c
